a float conversion

+++ mollib/src/main/java/com/bammellab/mollib/GLSurfaceViewDisplayPdbFile.kt
@@ -1,17 +1,14 @@
 /*
- * Copyright (C) 2016-2018 James Andreas
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License
+ *  Copyright 2020 Bammellab / James Andreas
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License
  */
 
 @file:Suppress("unused", "unused_variable", "unused_parameter", "deprecation")
@@ -270,10 +267,10 @@ class GLSurfaceViewDisplayPdbFile : GLSurfaceView {
      * Calculate the degree to be rotated by.
      */
     private fun rotation(event: MotionEvent): Float {
-        val deltaX = (event.getX(0) - event.getX(1)).toDouble()
-        val deltaY = (event.getY(0) - event.getY(1)).toDouble()
+        val deltaX = (event.getX(0) - event.getX(1)).toFloat()
+        val deltaY = (event.getY(0) - event.getY(1)).toFloat()
         val radians = atan2(deltaY, deltaX)
-        return Math.toDegrees(radians).toFloat()
+        return Math.toDegrees(radians.toDouble()).toFloat()
     }
     
     fun setRenderer(rendererIn: RendererDisplayPdbFile, densityIn: Int) {
